fetch/metadata/generated/element-video-poster.sub.tentative.html

<!DOCTYPE html>
<!--
This test was procedurally generated. Please do not modify it directly.
Sources:
- fetch/metadata/tools/fetch-metadata.conf.yml
- fetch/metadata/tools/templates/element-video-poster.sub.html
-->
<html lang="en">
  <meta charset="utf-8">
  <title>HTTP headers on request for HTML "video" element "poster"</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <body>
  <script>
  'use strict';

  const params = {
    body: `
      <svg xmlns="http://www.w3.org/2000/svg" width="123" height="123">
        <rect fill="lime" width="123" height="123"/>
      </svg>
    `,
    mime: 'image/svg+xml'
  };

  function induceRequest(t, url) {
    var video = document.createElement('video');
    video.setAttribute('poster', url);
    document.body.appendChild(video);

    const poll = () => {
      if (video.clientWidth === 123) {
        return;
      }

      return new Promise((resolve) => t.step_timeout(resolve, 0))
        .then(poll);
    };
    t.add_cleanup(() => video.remove());

    return poll();
  }

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(t, makeRequestURL(key, ['httpOrigin'], params))
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-frame-ancestors');
      });
  }, 'sec-fetch-frame-ancestors - Not sent to non-trustworthy same-origin destination');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(t, makeRequestURL(key, ['httpSameSite'], params))
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-frame-ancestors');
      });
  }, 'sec-fetch-frame-ancestors - Not sent to non-trustworthy same-site destination');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(t, makeRequestURL(key, ['httpCrossSite'], params))
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-frame-ancestors');
      });
  }, 'sec-fetch-frame-ancestors - Not sent to non-trustworthy cross-site destination');
  </script>
  </body>
</html>
